refactor: reorganize files into ra, utilities, and hooks
