+++ src/components/atoms/TwitchCCLSelect.tsx
@@ -3,7 +3,7 @@ import {
   type ContentClassificationLabel,
   type ContentClassificationLabelDefinition,
   getContentClassificationLabels,
-} from '@/twitch';
+} from '@/utilities/twitch';
 import Box from '@mui/material/Box';
 import Chip from '@mui/material/Chip';
 import FormControl from '@mui/material/FormControl';
