+++ src/components/atoms/TwitchCategoryAutocomplete.tsx
@@ -1,5 +1,5 @@
 import type { Profile } from '@/hooks/useProfile';
-import { type TwitchCategory, searchCategories } from '@/twitch';
+import { type TwitchCategory, searchCategories } from '@/utilities/twitch';
 import Alert from '@mui/material/Alert';
 import Autocomplete from '@mui/material/Autocomplete';
 import Box from '@mui/material/Box';
