+++ src/components/molecules/StreamInfoEditor.tsx
@@ -8,7 +8,7 @@ import {
   type GetChannelInformationResponse,
   getChannelInformation,
   modifyChannelInformation,
-} from '@/twitch';
+} from '@/utilities/twitch';
 import Alert from '@mui/material/Alert';
 import Box from '@mui/material/Box';
 import Button from '@mui/material/Button';
