+++ src/ra/dataProvider/restDataProvider.ts
@@ -1,4 +1,4 @@
-import { userManager } from '@/auth';
+import { userManager } from '@/utilities/auth';
 import type { DataProvider, Identifier } from 'react-admin';
 import { HttpError } from 'react-admin';
 import * as cursorPaginationCache from './cursorPaginationCache';
